moved Chip8 into own struct and module
